// src/utils/type.ts
export type PokeTypeColor =
  | "TypeGrass"
  | "TypeWater"
  | "TypePoison"
  | "TypeGround"
  | "TypeDark"
  | "TypeFire"
  | "TypeFlying"
  | "TypeBug"
  | "TypeElectric"
  | "TypeFighting"
  | "TypeRock"
  | "TypeSteel"
  | "TypeIce"
  | "TypeNormal"
  | "TypeFairy"
  | "TypeGhost"
  | "TypeDragon"
  | "TypePsychic";